Add y/n prompt to "delete_all" scripts.

+++ bash/delete_all_holdings.sh
@@ -3,5 +3,14 @@
 TMP='./.okapi'
 OKAPI=`cat ${TMP}/url`
 TOKEN=`cat ${TMP}/token`
+TENANT=`cat ${TMP}/tenant`
 
-curl --http1.1 -w '\n' -X DELETE "${OKAPI}/holdings-storage/holdings" -H "x-okapi-token: ${TOKEN}"
+read -n 1 -p "Are you sure you want to delete all holdings for tenant \"${TENANT}?\" (y/n): " YN
+echo
+if [ $YN == y ]
+  then
+    echo "Deleting all holdings..."
+    curl --http1.1 -w '\n' -X DELETE "${OKAPI}/holdings-storage/holdings" -H "x-okapi-token: ${TOKEN}"
+  else
+    echo "Aborting process..."
+fi
